backend/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
supabase==2.11.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Pydantic for data validation
pydantic==2.10.5
pydantic-settings==2.7.0

# AI Integration
google-generativeai==0.8.3

# Video Processing
ffmpeg-python==0.2.0

# HTTP Client
httpx==0.28.1

# Utilities
python-dateutil==2.9.0
